dom/media/tests/mochitest/test_peerConnection_bug822674.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=822674
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 822674</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="head.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript">

  /** Test for Bug 822674 **/
  runTest(function() {
    var pc = new mozRTCPeerConnection();
    pc.thereIsNeverGoingToBeAPropertyWithThisNameOnThisInterface = 1;
    is(pc.thereIsNeverGoingToBeAPropertyWithThisNameOnThisInterface, 1,
       "Can set expandos on an RTCPeerConnection")
    SimpleTest.finish();
  }, true);
  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=822674">Mozilla Bug 822674</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
